agents/mcp-server/pyproject.toml
[project]
name = "todo-mcp-server"
version = "0.1.0"
description = "MCP Server exposing task CRUD tools for AI agents"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.7,<2.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
